Split functions over smaller, more manageable, source files.
- No logical changes.
- Public components at top.
- Rename parse_module_declaration_ansi to module_declaration_ansi.
- Rename parse_module_declaration_nonansi to module_declaration_nonansi.
