+++ src/lib.rs
@@ -1,1 +1,4 @@
 pub mod structures;
+pub mod sv_misc;
+pub mod sv_module;
+pub mod sv_port;
