+++ src/main.rs
@@ -8,11 +8,9 @@ use std::fs::File;
 use std::io::{BufWriter, Write};
 use std::path::{Path, PathBuf};
 use std::{env, process};
-use sv_parser::{parse_sv, unwrap_node, Define, DefineText, NodeEvent, RefNode, SyntaxTree};
-use svdata::structures::{
-    SvData, SvDataKind, SvDataType, SvModuleDeclaration, SvNetType, SvParameter, SvPort,
-    SvPortDirection, SvSignedness,
-};
+use sv_parser::{parse_sv, Define, DefineText, NodeEvent, RefNode, SyntaxTree};
+use svdata::structures::SvData;
+use svdata::sv_module::{module_declaration_ansi, module_declaration_nonansi};
 use verilog_filelist_parser;
 
 #[derive(Debug, Parser)]
@@ -204,321 +202,16 @@ fn sv_to_structure(syntax_tree: &SyntaxTree, filepath: &str, svdata: &mut SvData
         if enter_not_leave {
             match node {
                 RefNode::ModuleDeclarationAnsi(_) => {
-                    let d = parse_module_declaration_ansi(node, &syntax_tree, filepath);
+                    let d = module_declaration_ansi(node, &syntax_tree, filepath);
                     svdata.modules.push(d.clone());
                 }
                 RefNode::ModuleDeclarationNonansi(_) => {
-                    let _d = parse_module_declaration_nonansi(node, &syntax_tree, filepath);
+                    let _d = module_declaration_nonansi(node, &syntax_tree, filepath);
                 }
                 _ => (),
             }
         }
     }
-}
-
-fn identifier(parent: RefNode, syntax_tree: &SyntaxTree) -> Option<String> {
-    let id = match unwrap_node!(parent, SimpleIdentifier, EscapedIdentifier) {
-        Some(RefNode::SimpleIdentifier(x)) => Some(x.nodes.0),
-        Some(RefNode::EscapedIdentifier(x)) => Some(x.nodes.0),
-        _ => None,
-    };
-
-    match id {
-        Some(x) => Some(syntax_tree.get_str(&x).unwrap().to_string()),
-        _ => None,
-    }
-}
-
-fn keyword(parent: RefNode, syntax_tree: &SyntaxTree) -> Option<String> {
-    let id = match unwrap_node!(parent, Keyword) {
-        Some(RefNode::Keyword(x)) => Some(x.nodes.0),
-
-        _ => None,
-    };
-
-    match id {
-        Some(x) => Some(syntax_tree.get_str(&x).unwrap().to_string()),
-        _ => None,
-    }
-}
-
-fn module_identifier(node: RefNode, syntax_tree: &SyntaxTree) -> Option<String> {
-    let id = unwrap_node!(node, ModuleIdentifier).unwrap();
-    identifier(id, &syntax_tree)
-}
-
-fn parse_module_declaration_ansi(
-    m: RefNode,
-    syntax_tree: &SyntaxTree,
-    filepath: &str,
-) -> SvModuleDeclaration {
-    let mut ret = SvModuleDeclaration {
-        identifier: module_identifier(m.clone(), syntax_tree).unwrap(),
-        parameters: Vec::new(),
-        ports: Vec::new(),
-        filepath: String::from(filepath),
-    };
-
-    let mut prev_port: Option<SvPort> = None;
-
-    for node in m {
-        match node {
-            RefNode::ParameterDeclarationParam(p) => ret
-                .parameters
-                .push(parse_module_declaration_parameter(p, syntax_tree)),
-            RefNode::AnsiPortDeclaration(p) => {
-                let parsed_port: SvPort =
-                    parse_module_declaration_port_ansi(p, syntax_tree, &prev_port.clone());
-                ret.ports.push(parsed_port.clone());
-                prev_port = Some(parsed_port.clone());
-            }
-            _ => (),
-        }
-    }
-    ret
-}
-
-fn parse_module_declaration_nonansi(
-    _m: RefNode,
-    _syntax_tree: &SyntaxTree,
-    _filepath: &str,
-) -> SvModuleDeclaration {
-    let ret = SvModuleDeclaration {
-        identifier: module_identifier(_m, _syntax_tree).unwrap(),
-        parameters: Vec::new(),
-        ports: Vec::new(),
-        filepath: String::from(_filepath),
-    };
-    // TODO
-    ret
-}
-
-fn parse_module_declaration_parameter(
-    p: &sv_parser::ParameterDeclarationParam,
-    _syntax_tree: &SyntaxTree,
-) -> SvParameter {
-    println!("parameter={:?}", p);
-    SvParameter {
-        identifier: String::from("foo"),
-        datatype: String::from("bar"),
-    }
-}
-
-fn port_identifier(node: &sv_parser::AnsiPortDeclaration, syntax_tree: &SyntaxTree) -> String {
-    let id = unwrap_node!(node, PortIdentifier).unwrap();
-    identifier(id, &syntax_tree).unwrap()
-}
-
-fn port_direction_ansi(
-    node: &sv_parser::AnsiPortDeclaration,
-    prev_port: &Option<SvPort>,
-) -> SvPortDirection {
-    let dir = unwrap_node!(node, PortDirection);
-    match dir {
-        Some(RefNode::PortDirection(sv_parser::PortDirection::Inout(_))) => SvPortDirection::Inout,
-        Some(RefNode::PortDirection(sv_parser::PortDirection::Input(_))) => SvPortDirection::Input,
-        Some(RefNode::PortDirection(sv_parser::PortDirection::Output(_))) => {
-            SvPortDirection::Output
-        }
-        Some(RefNode::PortDirection(sv_parser::PortDirection::Ref(_))) => SvPortDirection::Ref,
-        _ => match prev_port {
-            Some(_) => prev_port.clone().unwrap().direction,
-            None => SvPortDirection::Inout,
-        },
-    }
-}
-
-fn port_datakind_ansi(nettype: &Option<SvNetType>) -> SvDataKind {
-    match nettype {
-        None => SvDataKind::Variable,
-
-        Some(_) => SvDataKind::Net,
-    }
-}
-
-fn port_datatype_ansi(
-    node: &sv_parser::AnsiPortDeclaration,
-    syntax_tree: &SyntaxTree,
-) -> SvDataType {
-    let datatype = unwrap_node!(
-        node,
-        IntegerVectorType,
-        IntegerAtomType,
-        NonIntegerType,
-        ClassType,
-        TypeReference
-    );
-    match datatype {
-        Some(RefNode::IntegerVectorType(sv_parser::IntegerVectorType::Logic(_))) => {
-            SvDataType::Logic
-        }
-        Some(RefNode::IntegerVectorType(sv_parser::IntegerVectorType::Reg(_))) => SvDataType::Reg,
-        Some(RefNode::IntegerVectorType(sv_parser::IntegerVectorType::Bit(_))) => SvDataType::Bit,
-        Some(RefNode::IntegerAtomType(sv_parser::IntegerAtomType::Byte(_))) => SvDataType::Byte,
-        Some(RefNode::IntegerAtomType(sv_parser::IntegerAtomType::Shortint(_))) => {
-            SvDataType::Shortint
-        }
-        Some(RefNode::IntegerAtomType(sv_parser::IntegerAtomType::Int(_))) => SvDataType::Int,
-        Some(RefNode::IntegerAtomType(sv_parser::IntegerAtomType::Longint(_))) => {
-            SvDataType::Longint
-        }
-        Some(RefNode::IntegerAtomType(sv_parser::IntegerAtomType::Integer(_))) => {
-            SvDataType::Integer
-        }
-        Some(RefNode::IntegerAtomType(sv_parser::IntegerAtomType::Time(_))) => SvDataType::Time,
-        Some(RefNode::NonIntegerType(sv_parser::NonIntegerType::Shortreal(_))) => {
-            SvDataType::Shortreal
-        }
-        Some(RefNode::NonIntegerType(sv_parser::NonIntegerType::Realtime(_))) => {
-            SvDataType::Realtime
-        }
-        Some(RefNode::NonIntegerType(sv_parser::NonIntegerType::Real(_))) => SvDataType::Real,
-        Some(RefNode::ClassType(_)) => SvDataType::Class,
-        Some(RefNode::TypeReference(_)) => SvDataType::TypeRef,
-        _ => match unwrap_node!(node, DataType) {
-            Some(x) => match keyword(x, syntax_tree) {
-                Some(x) => {
-                    if x == "string" {
-                        return SvDataType::String;
-                    } else {
-                        println!("{}", x);
-                        unreachable!();
-                    }
-                }
-
-                _ => unreachable!(),
-            },
-            _ => return SvDataType::Logic,
-        },
-    }
-}
-
-fn port_nettype_ansi(
-    m: &sv_parser::AnsiPortDeclaration,
-    direction: &SvPortDirection,
-) -> Option<SvNetType> {
-    let nettype = unwrap_node!(m, AnsiPortDeclarationVariable, AnsiPortDeclarationNet);
-    match nettype {
-        Some(RefNode::AnsiPortDeclarationVariable(_)) => {
-            match unwrap_node!(m, PortDirection, DataType, Signing) {
-                Some(_) => return None,
-                _ => return Some(SvNetType::Wire),
-            }
-        }
-
-        Some(RefNode::AnsiPortDeclarationNet(x)) => {
-            let dir = unwrap_node!(x, NetType);
-
-            match dir {
-                // "Var" token was not found
-                Some(RefNode::NetType(sv_parser::NetType::Supply0(_))) => {
-                    return Some(SvNetType::Supply0)
-                }
-                Some(RefNode::NetType(sv_parser::NetType::Supply1(_))) => {
-                    return Some(SvNetType::Supply1)
-                }
-                Some(RefNode::NetType(sv_parser::NetType::Triand(_))) => {
-                    return Some(SvNetType::Triand)
-                }
-                Some(RefNode::NetType(sv_parser::NetType::Trior(_))) => {
-                    return Some(SvNetType::Trior)
-                }
-                Some(RefNode::NetType(sv_parser::NetType::Trireg(_))) => {
-                    return Some(SvNetType::Trireg)
-                }
-                Some(RefNode::NetType(sv_parser::NetType::Tri0(_))) => {
-                    return Some(SvNetType::Tri0)
-                }
-                Some(RefNode::NetType(sv_parser::NetType::Tri1(_))) => {
-                    return Some(SvNetType::Tri1)
-                }
-                Some(RefNode::NetType(sv_parser::NetType::Tri(_))) => return Some(SvNetType::Tri),
-                Some(RefNode::NetType(sv_parser::NetType::Uwire(_))) => {
-                    return Some(SvNetType::Uwire)
-                }
-                Some(RefNode::NetType(sv_parser::NetType::Wire(_))) => {
-                    return Some(SvNetType::Wire)
-                }
-                Some(RefNode::NetType(sv_parser::NetType::Wand(_))) => {
-                    return Some(SvNetType::Wand)
-                }
-                Some(RefNode::NetType(sv_parser::NetType::Wor(_))) => return Some(SvNetType::Wor),
-
-                _ => match direction {
-                    SvPortDirection::Inout | SvPortDirection::Input => {
-                        return Some(SvNetType::Wire);
-                    }
-                    SvPortDirection::Output => match unwrap_node!(m, DataType) {
-                        Some(_) => return None,
-                        _ => return Some(SvNetType::Wire),
-                    },
-
-                    SvPortDirection::Ref => {
-                        return None;
-                    }
-
-                    _ => unreachable!(),
-                },
-            }
-        }
-
-        _ => unreachable!(),
-    }
-}
-
-fn port_signedness_ansi(m: &sv_parser::AnsiPortDeclaration) -> SvSignedness {
-    let signedness = unwrap_node!(m, Signing);
-    match signedness {
-        Some(RefNode::Signing(sv_parser::Signing::Signed(_))) => SvSignedness::Signed,
-        Some(RefNode::Signing(sv_parser::Signing::Unsigned(_))) => SvSignedness::Unsigned,
-        _ => SvSignedness::Unsigned,
-    }
-}
-
-fn port_check_inheritance_ansi(
-    m: &sv_parser::AnsiPortDeclaration,
-    prev_port: &Option<SvPort>,
-) -> bool {
-    let datatype = unwrap_node!(m, DataType, Signing, NetType, VarDataType, PortDirection);
-
-    match prev_port {
-        Some(_) => match datatype {
-            Some(_) => false,
-            _ => true,
-        },
-        None => false,
-    }
-}
-
-fn parse_module_declaration_port_ansi(
-    p: &sv_parser::AnsiPortDeclaration,
-    syntax_tree: &SyntaxTree,
-    prev_port: &Option<SvPort>,
-) -> SvPort {
-    let inherit = port_check_inheritance_ansi(p, prev_port);
-    let ret: SvPort;
-
-    if inherit == false {
-        ret = SvPort {
-            identifier: port_identifier(p, syntax_tree),
-            direction: port_direction_ansi(p, prev_port),
-            nettype: port_nettype_ansi(p, &port_direction_ansi(p, prev_port)),
-            datakind: port_datakind_ansi(&port_nettype_ansi(p, &port_direction_ansi(p, prev_port))),
-            datatype: port_datatype_ansi(p, syntax_tree),
-            signedness: port_signedness_ansi(p),
-        }
-    } else {
-        ret = SvPort {
-            identifier: port_identifier(p, syntax_tree),
-            direction: prev_port.clone().unwrap().direction,
-            nettype: prev_port.clone().unwrap().nettype,
-            datakind: prev_port.clone().unwrap().datakind,
-            datatype: prev_port.clone().unwrap().datatype,
-            signedness: prev_port.clone().unwrap().signedness,
-        };
-    }
-
-    return ret;
 }
 
 #[cfg(test)]
